+++ packages/browser/playwright/mocked/session-recording/session-recording-ingestion-controls.spec.ts
@@ -4,7 +4,10 @@ import { assertThatRecordingStarted, pollUntilEventCaptured } from '../utils/eve
 
 const startOptions = {
     options: {
-        session_recording: {},
+        session_recording: {
+            // not the default but makes for easier test assertions
+            compress_events: false,
+        },
         opt_out_capturing_by_default: true,
     },
     flagsResponseOverrides: {