+++ packages/browser/playwright/mocked/session-recording/session-recording-network-recorder.spec.ts
@@ -68,7 +68,10 @@ test.beforeEach(async ({ context }) => {
                     await start(
                         {
                             options: {
-                                session_recording: {},
+                                session_recording: {
+                                    // not the default but makes for easier test assertions
+                                    compress_events: false,
+                                },
                             },
                             flagsResponseOverrides: {
                                 sessionRecording: {