+++ packages/browser/src/__tests__/extensions/replay/lazy-sessionrecording.test.ts
@@ -164,6 +164,7 @@ const originalLocation = window!.location
 
 function fakeNavigateTo(href: string) {
     delete (window as any).location
+    // @ts-expect-error this is a test, it's safe to write to location like this
     window!.location = { href } as Location
 }
 
@@ -216,6 +217,8 @@ describe('Lazy SessionRecording', () => {
             autocapture: false, // Assert that session recording works even if `autocapture = false`
             session_recording: {
                 maskAllInputs: false,
+                // not the default but makes for easier test assertions
+                compress_events: false,
             },
             persistence: 'memory',
         } as unknown as PostHogConfig
@@ -288,6 +291,7 @@ describe('Lazy SessionRecording', () => {
     })
 
     afterEach(() => {
+        // @ts-expect-error this is a test, it's safe to write to location like this
         window!.location = originalLocation
     })
 
@@ -1174,29 +1178,6 @@ describe('Lazy SessionRecording', () => {
                 )
             })
 
-            it('does not compress small full snapshot data', () => {
-                _emit(createFullSnapshot({ data: { content: 'small' } }))
-                sessionRecording['_lazyLoadedSessionRecording']['_flushBuffer']()
-
-                expect(posthog.capture).toHaveBeenCalledWith(
-                    '$snapshot',
-                    {
-                        $snapshot_data: [
-                            {
-                                data: { content: 'small' },
-                                type: 2,
-                            },
-                        ],
-                        $session_id: sessionId,
-                        $snapshot_bytes: expect.any(Number),
-                        $window_id: 'windowId',
-                        $lib: 'web',
-                        $lib_version: '0.0.1',
-                    },
-                    captureOptions
-                )
-            })
-
             it('compresses incremental snapshot mutation data', () => {
                 _emit(createIncrementalMutationEvent({ texts: [Array(30).fill(uuidv7()).join('')] }))
                 sessionRecording['_lazyLoadedSessionRecording']['_flushBuffer']()
@@ -1638,7 +1619,6 @@ describe('Lazy SessionRecording', () => {
         })
 
         it('can emit when there are circular references', () => {
-            posthog.config.session_recording.compress_events = false
             sessionRecording.onRemoteConfig(makeFlagsResponse({ sessionRecording: { endpoint: '/s/' } }))
             sessionRecording.onRemoteConfig(
                 makeFlagsResponse({