+++ packages/browser/src/extensions/replay/triggerMatching.ts
@@ -137,9 +137,17 @@ export class URLTriggerMatching implements TriggerStatusMatching {
 
     onConfig(config: ReplayConfigType) {
         this._urlTriggers =
-            (isEagerLoadedConfig(config) ? config.sessionRecording?.urlTriggers : config?.urlTriggers) || []
+            (isEagerLoadedConfig(config)
+                ? isObject(config.sessionRecording)
+                    ? config.sessionRecording?.urlTriggers
+                    : []
+                : config?.urlTriggers) || []
         this._urlBlocklist =
-            (isEagerLoadedConfig(config) ? config.sessionRecording?.urlBlocklist : config?.urlBlocklist) || []
+            (isEagerLoadedConfig(config)
+                ? isObject(config.sessionRecording)
+                    ? config.sessionRecording?.urlBlocklist
+                    : []
+                : config?.urlBlocklist) || []
     }
 
     /**
@@ -225,7 +233,11 @@ export class LinkedFlagMatching implements TriggerStatusMatching {
 
     onConfig(config: ReplayConfigType, onStarted: (flag: string, variant: string | null) => void) {
         this.linkedFlag =
-            (isEagerLoadedConfig(config) ? config.sessionRecording?.linkedFlag : config?.linkedFlag) || null
+            (isEagerLoadedConfig(config)
+                ? isObject(config.sessionRecording)
+                    ? config.sessionRecording?.linkedFlag
+                    : null
+                : config?.linkedFlag) || null
 
         if (!isNullish(this.linkedFlag) && !this.linkedFlagSeen) {
             const linkedFlag = isString(this.linkedFlag) ? this.linkedFlag : this.linkedFlag.flag
@@ -271,7 +283,11 @@ export class EventTriggerMatching implements TriggerStatusMatching {
 
     onConfig(config: ReplayConfigType) {
         this._eventTriggers =
-            (isEagerLoadedConfig(config) ? config.sessionRecording?.eventTriggers : config?.eventTriggers) || []
+            (isEagerLoadedConfig(config)
+                ? isObject(config.sessionRecording)
+                    ? config.sessionRecording?.eventTriggers
+                    : []
+                : config?.eventTriggers) || []
     }
 
     /**